design: 디테일 페이지 스켈레톤 수정
