+++ client/src/components/community/skeleton/SkeletonMapContainer.tsx
@@ -1,10 +1,7 @@
 import { styled } from 'styled-components';
 
-import SkeletonLocationContainer from './SkeletonLocationContainer';
-
 import cssToken from '../../../styles/cssToken';
 import { FlexDiv } from '../../../styles/styles';
-import { TitleDiv } from '../../skeleton/SkeletonContentCard';
 import SkeletonMap from '../../skeleton/SkeletonMap';
 
 const MapContainerDiv = styled(FlexDiv)`
@@ -23,27 +20,14 @@ const MapContainerDiv = styled(FlexDiv)`
     .skeletonMap {
       margin-left: 0px;
       flex: 1;
-      order: -1;
-      height: 60vh;
+      height: 50vh;
     }
   }
-`;
-
-const ScheduleDiv = styled(FlexDiv)`
-  flex-direction: column;
-  flex: 1;
-  height: 60vh;
-  overflow: auto;
-  row-gap: ${cssToken.SPACING['gap-12']};
 `;
 
 const SkeletonMapContainer = () => {
   return (
     <MapContainerDiv className="skeletonScheduleDiv">
-      <ScheduleDiv>
-        <TitleDiv className="skeletonMapTitle" />
-        <SkeletonLocationContainer length={3} />
-      </ScheduleDiv>
       <SkeletonMap />
     </MapContainerDiv>
   );
