+++ client/src/components/skeleton/SkeletonMap.tsx
@@ -7,7 +7,7 @@ const MapDiv = styled(SkeletonDiv)`
   border-radius: 0rem;
   margin-left: ${cssToken.SPACING['gap-24']};
   flex: 3;
-  height: 60vh;
+  height: 80vh;
 `;
 const SkeletonMap = () => {
   return <MapDiv className="skeletonMap" />;
